requirements.txt
langchain==0.3.27
langchain_community==0.3.29
langchain_core==0.3.76
langchain_google_genai==2.1.10
langchain_huggingface==0.3.1
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.49.1
faiss-cpu==1.12.0
sentence-transformers==5.1.0
